Introduce a more compact format to describe tunes
A simple tune would look as follows:

C4-4 D4-8 E4-8 -4 D4-8 C4-4.

Each note is delimited by a space, and is described as a note-duration
pair. The pair is separated by a hyphen.

Ties are denoted by skipping the note name.
Dotted notes are denoted by a dot after the duration.

The duration represents the fractional duration, to make it easier to
translate written music to this format. For example, -8represents and
1/8th note (as opposed to 8 beats).